+++ changelog.rst
@@ -4,6 +4,11 @@ Compysition changelog
 Version
 1.3.1
 
+- Extracted accept header logic from HTTPServer into an AcceptPlugin.
+    - No longer ignores Accept header altogether when an unknown mime-type is encountered.
+    - Implements route based Accept definitions.
+    - No longer selects target response Content-Type prior to service processing.  Now performs narrowing functions upon receiving a request and selects target response Content-Type immediately prior to send response.  Narrowed mime-types are stored in responders dict to mimize event size.  Selection now prioritizes matching incoming mime-type, service desired mime-type, current event type's corresponding mime-type, and defaults to first available mime-type from narrowing process.
+    - Implements new response object _CompysitionHTTPError which is used to pass a request's matching route to HTTPServer's optional Bottle error handler.  Allowing for route Accept type definitions to be applied to these Bottle handled errors.
 - Added python2/python3/python3.5+ flags.
     - Used flags to implement version specific functionality (s.a. iterables, imports, etc.).
 - Made adjustments to remove warnings.